graph work
